extract mergeConditions function
