+++ packages/sdk/src/permissions/mergeFunctionPermissions.ts
@@ -1,5 +1,4 @@
-import { Condition, Operator, ParameterType } from "zodiac-roles-deployments"
-
+import { mergeConditions } from "./mergeConditions"
 import { coercePermission, targetId } from "./utils"
 
 import {
@@ -47,33 +46,3 @@ export const mergeFunctionPermissions = (permissions: Permission[]) =>
 
     return result
   }, [] as PermissionCoerced[])
-
-/**
- * @dev Merges two conditions using a logical OR, flattening nested OR conditions. If the conditions are equal, it will still create separate OR branches.
- * These will be pruned later in sanitizeCondition().
- */
-const mergeConditions = (
-  a: FunctionPermissionCoerced,
-  b: FunctionPermissionCoerced
-): Condition | undefined => {
-  if (!!a.condition !== !!b.condition) {
-    const id = targetId(a)
-    console.warn(
-      `Target ${id} is allowed with and without conditions. It will be allowed without conditions.`
-    )
-    return undefined
-  }
-
-  if (!a.condition || !b.condition) return undefined
-
-  const aBranches =
-    a.condition.operator === Operator.Or ? a.condition.children : [a.condition]
-  const bBranches =
-    b.condition.operator === Operator.Or ? b.condition.children : [b.condition]
-
-  return {
-    paramType: ParameterType.None,
-    operator: Operator.Or,
-    children: [...(aBranches || []), ...(bBranches || [])],
-  }
-}
